+++ doc/cppobj.Doxyfile
@@ -85,8 +85,8 @@ QUIET = YES
 # spaces.
 # Note: If this tag is empty the current directory is searched.
 
-INPUT = ../src/include/libpmemobj
-INPUT += ../src/include/libpmemobj/README.md
+INPUT = ../src/include/libpmemobj++
+INPUT += ../src/include/libpmemobj++/README.md
 
 # This tag can be used to specify the character encoding of the source files
 # that doxygen parses. Internally doxygen uses the UTF-8 encoding. Doxygen uses
