+++ src/Makefile
@@ -56,7 +56,7 @@ HEADERS_INSTALL = include/libpmem.h include/libvmem.h\
 		include/libpmemblk.h include/libpmemlog.h\
 		include/libvmmalloc.h
 
-CPP_HEADERS_DESTDIR = $(DESTDIR)$(includedir)/libpmemobj
+CPP_HEADERS_DESTDIR = $(DESTDIR)$(includedir)/libpmemobj++
 
 MAKE_PKG_CONFIG = $(TOP)/utils/make-pkg-config.sh
 PKG_CONFIG_DESTDIR = $(DESTDIR)$(pkgconfigdir)
@@ -140,7 +140,7 @@ endif
 
 ifeq ($(EXPERIMENTAL),y)
 install-cpp:
-	$(call install_recursive_filter,include/libpmemobj,*.hpp,0644,$(CPP_HEADERS_DESTDIR))
+	$(call install_recursive_filter,include/libpmemobj++,*.hpp,0644,$(CPP_HEADERS_DESTDIR))
 
 install: install-cpp
 
@@ -164,14 +164,14 @@ uninstall:
 # [!q] to skip windows/include/sys/queue.h
 cstyle:
 	$(STYLE_CHECK) check include/*.h
-	$(STYLE_CHECK) check include/libpmemobj/*.hpp
-	$(STYLE_CHECK) check include/libpmemobj/detail/*.hpp
+	$(STYLE_CHECK) check include/libpmemobj++/*.hpp
+	$(STYLE_CHECK) check include/libpmemobj++/detail/*.hpp
 	$(STYLE_CHECK) check windows/include/*.h
 	$(STYLE_CHECK) check windows/include/*/[!q]*.h
 
 format:
-	$(STYLE_CHECK) format include/libpmemobj/*.hpp
-	$(STYLE_CHECK) format include/libpmemobj/detail/*.hpp
+	$(STYLE_CHECK) format include/libpmemobj++/*.hpp
+	$(STYLE_CHECK) format include/libpmemobj++/detail/*.hpp
 
 cscope:
 	cscope -q -b $(SCOPEFILES)
