+++ src/include/libpmemobj++/condition_variable.hpp
@@ -42,8 +42,8 @@
 #include <condition_variable>
 
 #include "libpmemobj.h"
-#include "libpmemobj/detail/conversions.hpp"
-#include "libpmemobj/mutex.hpp"
+#include "libpmemobj++/detail/conversions.hpp"
+#include "libpmemobj++/mutex.hpp"
 
 namespace nvml
 {
