+++ src/include/libpmemobj++/detail/array_traits.hpp
