+++ src/include/libpmemobj++/detail/specialization.hpp
