+++ src/include/libpmemobj++/make_persistent_array_atomic.hpp
@@ -41,11 +41,11 @@
 #define PMEMOBJ_MAKE_PERSISTENT_ARRAY_ATOMIC_HPP
 
 #include "libpmemobj.h"
-#include "libpmemobj/detail/array_traits.hpp"
-#include "libpmemobj/detail/check_persistent_ptr_array.hpp"
-#include "libpmemobj/detail/common.hpp"
-#include "libpmemobj/detail/make_atomic_impl.hpp"
-#include "libpmemobj/detail/pexceptions.hpp"
+#include "libpmemobj++/detail/array_traits.hpp"
+#include "libpmemobj++/detail/check_persistent_ptr_array.hpp"
+#include "libpmemobj++/detail/common.hpp"
+#include "libpmemobj++/detail/make_atomic_impl.hpp"
+#include "libpmemobj++/detail/pexceptions.hpp"
 
 namespace nvml
 {
