+++ src/include/libpmemobj++/make_persistent_atomic.hpp
@@ -41,11 +41,11 @@
 #define PMEMOBJ_MAKE_PERSISTENT_ATOMIC_HPP
 
 #include "libpmemobj.h"
-#include "libpmemobj/detail/check_persistent_ptr_array.hpp"
-#include "libpmemobj/detail/common.hpp"
-#include "libpmemobj/detail/make_atomic_impl.hpp"
-#include "libpmemobj/detail/pexceptions.hpp"
-#include "libpmemobj/pool.hpp"
+#include "libpmemobj++/detail/check_persistent_ptr_array.hpp"
+#include "libpmemobj++/detail/common.hpp"
+#include "libpmemobj++/detail/make_atomic_impl.hpp"
+#include "libpmemobj++/detail/pexceptions.hpp"
+#include "libpmemobj++/pool.hpp"
 
 #include <tuple>
