+++ src/include/libpmemobj++/p.hpp
@@ -41,8 +41,8 @@
 #include "libpmemobj.h"
 #include <memory>
 
-#include "libpmemobj/detail/common.hpp"
-#include "libpmemobj/detail/specialization.hpp"
+#include "libpmemobj++/detail/common.hpp"
+#include "libpmemobj++/detail/specialization.hpp"
 
 namespace nvml
 {
