+++ src/include/libpmemobj++/persistent_ptr.hpp
@@ -43,9 +43,9 @@
 #include <ostream>
 
 #include "libpmemobj.h"
-#include "libpmemobj/detail/common.hpp"
-#include "libpmemobj/detail/specialization.hpp"
-#include "libpmemobj/pool.hpp"
+#include "libpmemobj++/detail/common.hpp"
+#include "libpmemobj++/detail/specialization.hpp"
+#include "libpmemobj++/pool.hpp"
 
 namespace nvml
 {
