+++ src/include/libpmemobj++/timed_mutex.hpp
@@ -41,7 +41,7 @@
 #include <chrono>
 
 #include "libpmemobj.h"
-#include "libpmemobj/detail/conversions.hpp"
+#include "libpmemobj++/detail/conversions.hpp"
 
 namespace nvml
 {
