+++ src/test/obj_cpp_cond_var_posix/obj_cpp_cond_var_posix.cpp
@@ -36,9 +36,9 @@
 
 #include "unittest.h"
 
-#include <libpmemobj/condition_variable.hpp>
-#include <libpmemobj/persistent_ptr.hpp>
-#include <libpmemobj/pool.hpp>
+#include <libpmemobj++/condition_variable.hpp>
+#include <libpmemobj++/persistent_ptr.hpp>
+#include <libpmemobj++/pool.hpp>
 
 #include <mutex>
 #include <vector>
