+++ src/test/obj_cpp_make_persistent_array/obj_cpp_make_persistent_array.cpp
@@ -36,11 +36,11 @@
 
 #include "unittest.h"
 
-#include <libpmemobj/make_persistent_array.hpp>
-#include <libpmemobj/p.hpp>
-#include <libpmemobj/persistent_ptr.hpp>
-#include <libpmemobj/pool.hpp>
-#include <libpmemobj/transaction.hpp>
+#include <libpmemobj++/make_persistent_array.hpp>
+#include <libpmemobj++/p.hpp>
+#include <libpmemobj++/persistent_ptr.hpp>
+#include <libpmemobj++/pool.hpp>
+#include <libpmemobj++/transaction.hpp>
 
 #define LAYOUT "cpp"
