+++ src/test/obj_cpp_make_persistent_array_atomic/obj_cpp_make_persistent_array_atomic.cpp
@@ -37,10 +37,10 @@
 
 #include "unittest.h"
 
-#include <libpmemobj/make_persistent_array_atomic.hpp>
-#include <libpmemobj/p.hpp>
-#include <libpmemobj/persistent_ptr.hpp>
-#include <libpmemobj/pool.hpp>
+#include <libpmemobj++/make_persistent_array_atomic.hpp>
+#include <libpmemobj++/p.hpp>
+#include <libpmemobj++/persistent_ptr.hpp>
+#include <libpmemobj++/pool.hpp>
 
 #define LAYOUT "cpp"
