+++ src/test/obj_cpp_mutex_posix/obj_cpp_mutex_posix.cpp
@@ -36,9 +36,9 @@
 
 #include "unittest.h"
 
-#include <libpmemobj/mutex.hpp>
-#include <libpmemobj/persistent_ptr.hpp>
-#include <libpmemobj/pool.hpp>
+#include <libpmemobj++/mutex.hpp>
+#include <libpmemobj++/persistent_ptr.hpp>
+#include <libpmemobj++/pool.hpp>
 
 #include <mutex>
