+++ src/test/obj_cpp_pool_primitives/obj_cpp_pool_primitives.cpp
@@ -41,9 +41,9 @@
 
 #include "unittest.h"
 
-#include <libpmemobj/p.hpp>
-#include <libpmemobj/persistent_ptr.hpp>
-#include <libpmemobj/pool.hpp>
+#include <libpmemobj++/p.hpp>
+#include <libpmemobj++/persistent_ptr.hpp>
+#include <libpmemobj++/pool.hpp>
 
 namespace nvobj = nvml::obj;
