+++ src/test/obj_cpp_ptr/obj_cpp_ptr.cpp
@@ -37,13 +37,13 @@
 
 #include "unittest.h"
 
-#include <libpmemobj/make_persistent.hpp>
-#include <libpmemobj/make_persistent_array_atomic.hpp>
-#include <libpmemobj/make_persistent_atomic.hpp>
-#include <libpmemobj/p.hpp>
-#include <libpmemobj/persistent_ptr.hpp>
-#include <libpmemobj/pool.hpp>
-#include <libpmemobj/transaction.hpp>
+#include <libpmemobj++/make_persistent.hpp>
+#include <libpmemobj++/make_persistent_array_atomic.hpp>
+#include <libpmemobj++/make_persistent_atomic.hpp>
+#include <libpmemobj++/p.hpp>
+#include <libpmemobj++/persistent_ptr.hpp>
+#include <libpmemobj++/pool.hpp>
+#include <libpmemobj++/transaction.hpp>
 
 #define LAYOUT "cpp"
