+++ src/test/obj_cpp_ptr_arith/obj_cpp_ptr_arith.cpp
@@ -37,10 +37,10 @@
 
 #include "unittest.h"
 
-#include <libpmemobj/make_persistent_atomic.hpp>
-#include <libpmemobj/p.hpp>
-#include <libpmemobj/persistent_ptr.hpp>
-#include <libpmemobj/transaction.hpp>
+#include <libpmemobj++/make_persistent_atomic.hpp>
+#include <libpmemobj++/p.hpp>
+#include <libpmemobj++/persistent_ptr.hpp>
+#include <libpmemobj++/transaction.hpp>
 
 #include <sstream>
