+++ src/test/obj_cpp_shared_mutex/obj_cpp_shared_mutex.cpp
@@ -36,9 +36,9 @@
 
 #include "unittest.h"
 
-#include <libpmemobj/persistent_ptr.hpp>
-#include <libpmemobj/pool.hpp>
-#include <libpmemobj/shared_mutex.hpp>
+#include <libpmemobj++/persistent_ptr.hpp>
+#include <libpmemobj++/pool.hpp>
+#include <libpmemobj++/shared_mutex.hpp>
 
 #include <mutex>
 #include <thread>
