+++ src/test/obj_cpp_timed_mtx/obj_cpp_timed_mtx.cpp
@@ -36,9 +36,9 @@
 
 #include "unittest.h"
 
-#include <libpmemobj/persistent_ptr.hpp>
-#include <libpmemobj/pool.hpp>
-#include <libpmemobj/timed_mutex.hpp>
+#include <libpmemobj++/persistent_ptr.hpp>
+#include <libpmemobj++/pool.hpp>
+#include <libpmemobj++/timed_mutex.hpp>
 
 #include <mutex>
 #include <thread>
